+++ YoutubeCloneApp/Utilities/Extensions/UIResponder+Ext.swift
@@ -1,0 +1,16 @@
+import UIKit
+
+extension UIResponder {
+    private weak static var _currentFirstResponder: UIResponder?
+
+    static var currentFirstResponder: UIResponder? {
+        _currentFirstResponder = nil
+        UIApplication.shared.sendAction(#selector(findFirstResponder(sender:)), to: nil, from: nil, for: nil)
+        return _currentFirstResponder
+    }
+
+    @objc private func findFirstResponder(sender: Any) {
+        UIResponder._currentFirstResponder = self
+    }
+}
+
