+++ YoutubeCloneApp/Utilities/Extensions/UIViewController+Ext.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+extension UIViewController {
+    var name: String { String(describing: type(of: self)) }
+}
